third_party/WebKit/LayoutTests/media/video-not-paused-while-looping-expected.txt

EVENT(seeked)
EVENT(seeking)
This is a testharness.js-based test.
PASS video is not paused while looping 
Harness: the test ran to completion.
